avatar/api/requirements.txt
djangorestframework
